+++ src/components/AppElements/components/CodeEditor.vue
@@ -17,6 +17,7 @@
 			<v-tab
 				v-show="
 					route !== '/htmlgenerator' &&
+						route !== '/htmlimportplugin' &&
 						route !== '/htmlimport' &&
 						route !== '/sdk' &&
 						route !== '/gallery'
@@ -30,6 +31,7 @@
 					|| route === '/gallery'
 					|| route === '/variableeditor'
 					|| route === '/emailthumbnail'
+					|| route === '/htmlimportplugin'
 				"
 			>
 				Hooks
@@ -40,6 +42,7 @@
 					|| route === '/gallery'
 					|| route === '/variableeditor'
 					|| route === '/emailthumbnail'
+					|| route === '/htmlimportplugin'
 				"
 			>
 				Methods
@@ -56,7 +59,6 @@
 			<v-tab v-show="route === '/htmlgenerator'">
 				Output HTML
 			</v-tab>
-
 			<v-tab v-show="route === '/htmlimport'">
 				Input HTML
 			</v-tab>
@@ -219,6 +221,7 @@ import megaGalleryCodeGenerator from "./CodeEditor/codeGenerators/megaGalleryCod
 import variableEditorCodeGenerator from "./CodeEditor/codeGenerators/variableEditorCodeGenerator";
 import blockLibrariesCodeGenerator from "./CodeEditor/codeGenerators/blockLibrariesCodeGenerator";
 import documentCodeGenerator from "./CodeEditor/codeGenerators/documentCodeGenerator";
+import htmlImportPluginCodeGenerator from "./CodeEditor/codeGenerators/htmlImportPluginCodeGenerator";
 import htmlGeneratorCodeGenerator from "./CodeEditor/codeGenerators/htmlGeneratorCodeGenerator";
 import htmlImportCodeGenerator from "./CodeEditor/codeGenerators/htmlImportCodeGenerator";
 import dummyHtmlCodeGenerator from "./CodeEditor/codeGenerators/dummyHtmlCodeGenerator";
@@ -228,12 +231,14 @@ import variableEditorHooksGenerator from "./CodeEditor/hooks/variableEditorHooks
 import emailEditorHooksGenerator from "./CodeEditor/hooks/emailEditorHooks";
 import megaGalleryHooksGenerator from "./CodeEditor/hooks/megaGalleryHooks";
 import thumbnailHooksGenerator from "./CodeEditor/hooks/thumbnailHooks";
+import htmlImportPluginHooksGenerator from "./CodeEditor/hooks/htmlImportPluginHooks";
 
 import thumbnailMethodsGenerator from "./CodeEditor/methods/thumbnailMethods";
 import previewMethodsGenerator from "./CodeEditor/methods/previewMethods";
 import emailEditorMethodsGenerator from "./CodeEditor/methods/emailEditorMethods";
 import megaGalleryMethodsGenerator from "./CodeEditor/methods/megaGalleryMethods";
 import variableEditorMethodsGenerator from "./CodeEditor/methods/variableEditorMethods";
+import htmlImportPluginMethodsGenerator from "./CodeEditor/methods/htmlImportPluginMethods";
 
 import { mapGetters } from "vuex";
 
@@ -249,6 +254,7 @@ export default {
 		...mapGetters({ menus: "getMenu" }),
 		...mapGetters([
 			"getHtmlGeneratorConfigObject",
+			"getImportPluginConfigObject",
 			"getHtmlDocument",
 			"getDummyHtmlDocument",
 			"getDummyJSON",
@@ -323,6 +329,18 @@ export default {
 				this.$store.getters.getVariableEditorConfigObject,
 			);
 		},
+		// Html import plugin
+		htmlImportPluginCode() {
+			return htmlImportPluginCodeGenerator(this.getImportPluginConfigObject.settings);
+		},
+		htmlImportPluginHooks() {
+			return htmlImportPluginHooksGenerator();
+		},
+		htmlImportPluginMethods() {
+			return htmlImportPluginMethodsGenerator(
+				this.$store.getters.getImportPluginConfigObject.settings,
+			);
+		},
 		// Html generator
 		htmlGeneratorCode() {
 			return htmlGeneratorCodeGenerator(this.getHtmlGeneratorConfigObject);
@@ -353,8 +371,11 @@ export default {
 				return this.variableEditorCode;
 			} else if (this.$route.path === "/htmlgenerator") {
 				return this.htmlGeneratorCode;
-			} else if (this.$route.path === "/htmlimport") return this.htmlImportCode;
-			else return "//There is no code available";
+			} else if (this.$route.path === "/htmlimportplugin") {
+				return this.htmlImportPluginCode;
+			} else if (this.$route.path === "/htmlimport") {
+				return this.htmlImportCode;
+			} else return "//There is no code available";
 		},
 
 		hooks() {
@@ -363,6 +384,7 @@ export default {
 			else if (this.$route.path === "/gallery") return this.galleryHooks;
 			else if (this.$route.path === "/variableeditor") return this.variableEditorHooks;
 			else if (this.$route.path === "/emailthumbnail") return this.thumbnailHooks;
+			else if (this.$route.path === "/htmlimportplugin") return this.htmlImportPluginHooks;
 			else return "//There are no hooks available";
 		},
 		methods() {
@@ -371,6 +393,7 @@ export default {
 			else if (this.$route.path === "/gallery") return this.galleryMethods;
 			else if (this.$route.path === "/variableeditor") return this.variableEditorMethods;
 			else if (this.$route.path === "/emailthumbnail") return this.thumbnailMethods;
+			else if (this.$route.path === "/htmlimportplugin") return this.htmlImportPluginMethods;
 			else return "//There are no methods available";
 		},
 	},