+++ src/components/AppElements/store/menuData.js
@@ -183,6 +183,19 @@ export default {
 				],
 			},
 			{
+				title: "Email HTML Import Plugin",
+				icon: "code-json",
+				to: "htmlimportplugin",
+				children: [
+					{
+						title: "Header",
+						icon: "border-top-variant",
+						to: "#header",
+						codePropToMatch: "header",
+					},
+				],
+			},
+			{
 				title: "Email HTML Generator",
 				icon: "xml",
 				to: "htmlgenerator",
@@ -197,7 +210,7 @@ export default {
 			},
 			{
 				title: "Email HTML Import",
-				icon: "code-json",
+				icon: "code-tags",
 				to: "htmlimport",
 			},
 		],