+++ src/components/HtmlImport/view/HtmlImport.vue
@@ -1,31 +1,21 @@
 <template>
 	<div>
-		<PreviewButton
-			:button-text="'Open HTML import'"
-			:preview-button-visible="previewButtonVisible"
-			:is-inited="isInited"
-			@previewClick="openHtmlImport"
-		/>
 		<SectionObserver>
-			<div id="home" class="section">
+			<div class="section" id="home">
 				<Description
 					:title="'Email HTML Import'"
-					:doc-url="'https://chamaileon.io/sdk/v2/docs/email-html-importer/'"
+					:docUrl="'https://chamaileon.io/sdk/v2/docs/email-html-importer/'"
 					:image="'HtmlImportIllustration.svg'"
-					:button-text="/*'Open HTML import'*/null"
-					@showPreviewButton="showPreviewButton"
-					@previewClick="openHtmlImport"
+					:hideGetStarted="true"
 				>
 					<p>
-						Our HTML import plugin allows you to import an existing HTML email
-						to the editor and make changes to it. The output is a JSON object,
-						which the editor can process. Currently, there are no available
-						options for this plugin.
+						Our API allows you to send existing HTML email code to the backend, and convert it to a JSON object,
+						which can be processed in the editor. You can see examples under the right menu options
 					</p>
 				</Description>
 			</div>
 		</SectionObserver>
-		<Footer :previous="'Email HTML Generator'" :prev-to="'/htmlgenerator'" />
+		<Footer :previous="'Email HTML Generator'" :prevTo="'/htmlgenerator'" />
 	</div>
 </template>
 
@@ -33,57 +23,16 @@
 import SectionObserver from "../../AppElements/components/SectionObserver.vue";
 import Footer from "../../ViewUtilities/components/Footer.vue";
 import Description from "../../ViewUtilities/components/ViewDescription.vue";
-import PreviewButton from "../../AppElements/components/PreviewButton.vue";
-import { mapState } from "vuex";
 
 export default {
+	mounted() {
+		this.$store.dispatch("fetchDummyHtml");
+	},
+
 	components: {
 		SectionObserver,
 		Description,
 		Footer,
-		PreviewButton,
-	},
-	data() {
-		return {
-			previewButtonVisible: true,
-		};
-	},
-	computed: {
-		...mapState({
-			htmlImportInited: state => state.htmlImportInited,
-			sdkInited: state => state.sdkInited,
-		}),
-		isInited() {
-			if (this.sdkInited === true) {
-				return this.htmlImportInited;
-			}
-			return "pending";
-		},
-	},
-	watch: {
-		isInited: {
-			handler(v) {
-				if (v === false) {
-					this.$store.dispatch("initHtmlImport");
-				}
-			},
-			immediate: true,
-		},
-	},
-	mounted() {
-		this.$store.dispatch("fetchDummyHtml");
-	},
-	methods: {
-		async openHtmlImport() {
-			if (this.isInited === false) {
-				await this.$store.dispatch("initHtmlImport");
-			}
-			this.$chamaileon.htmlImport.show();
-			return;
-		},
-		showPreviewButton(isVisible) {
-			this.previewButtonVisible = isVisible;
-		},
 	},
 };
 </script>