+++ src/components/Lists/components/List6.vue
@@ -318,7 +318,7 @@ export default {
 			required: true,
 			validator(value) {
 				// The value must match one of these strings
-				return ["Editor", "Preview"].indexOf(value) !== -1;
+				return ["Editor", "Preview", "ImportPlugin"].indexOf(value) !== -1;
 			},
 		},
 	},
@@ -329,18 +329,19 @@ export default {
 
 		buttonsArr: {
 			get() {
-				return this.$store.state[this.section.toLowerCase() + "Config"].settings
+				return this.$store.state[this.section.charAt(0).toLowerCase() + this.section.slice(1) + "Config"].settings
 					.buttons.header;
 			},
 			set(value) {
 				this.$store.commit(`update${this.section}BtnOrder`, value);
 				this.updateEditorSettings();
 				this.updatePreviewSettings();
+				this.updateImportPluginSettings();
 			},
 		},
 		ddArrById: {
 			get(parentId) {
-				return this.$store.state[this.section.toLowerCase() + "Config"].settings
+				return this.$store.state[this.section.charAt(0).toLowerCase() + this.section.slice(1) + "Config"].settings
 					.buttons.header[parentId].items;
 			},
 			set(parentId, newArr) {
@@ -350,6 +351,7 @@ export default {
 				});
 				this.updateEditorSettings();
 				this.updatePreviewSettings();
+				this.updateImportPluginSettings();
 			},
 		},
 	},
@@ -357,6 +359,7 @@ export default {
 		...mapActions({
 			updateEditorSettings: "updateEditorSettings",
 			updatePreviewSettings: "updatePreviewSettings",
+			updateImportPluginSettings: "updateImportPluginSettings",
 		}),
 		updateLabel(val, index) {
 			this.$store.commit(`update${this.section}Btn`, {
@@ -365,6 +368,7 @@ export default {
 			});
 			this.updateEditorSettings();
 			this.updatePreviewSettings();
+			this.updateImportPluginSettings();
 		},
 		updateColor(val, index) {
 			this.$store.commit(`update${this.section}Btn`, {
@@ -373,6 +377,7 @@ export default {
 			});
 			this.updateEditorSettings();
 			this.updatePreviewSettings();
+			this.updateImportPluginSettings();
 		},
 		updateIcon(val, index) {
 			this.$store.commit(`update${this.section}Btn`, {
@@ -381,6 +386,7 @@ export default {
 			});
 			this.updateEditorSettings();
 			this.updatePreviewSettings();
+			this.updateImportPluginSettings();
 		},
 		updateStyle(val, index) {
 			this.$store.commit(`update${this.section}Btn`, {
@@ -389,11 +395,13 @@ export default {
 			});
 			this.updateEditorSettings();
 			this.updatePreviewSettings();
+			this.updateImportPluginSettings();
 		},
 		updateID(val, index) {
 			this.$store.commit(`update${this.section}Btn`, { index, id: val });
 			this.updateEditorSettings();
 			this.updatePreviewSettings();
+			this.updateImportPluginSettings();
 		},
 		updateDDID(val, parentIndex, index) {
 			this.$store.commit(`update${this.section}DropdownBtn`, {
@@ -402,11 +410,13 @@ export default {
 			});
 			this.updateEditorSettings();
 			this.updatePreviewSettings();
+			this.updateImportPluginSettings();
 		},
 		updateDDBtn(payload) {
 			this.$store.commit(`update${this.section}DropdownBtn`, payload);
 			this.updateEditorSettings();
 			this.updatePreviewSettings();
+			this.updateImportPluginSettings();
 		},
 		addDDBtn(payload) {
 			this.$store.commit(`add${this.section}DropdownBtn`, payload);
@@ -420,11 +430,13 @@ export default {
 			this.$store.commit(`update${this.section}DropdownBtnOrder`, payload);
 			this.updateEditorSettings();
 			this.updatePreviewSettings();
+			this.updateImportPluginSettings();
 		},
 		deleteBtn(payload) {
 			this.$store.commit(`remove${this.section}Btn`, payload);
 			this.updateEditorSettings();
 			this.updatePreviewSettings();
+			this.updateImportPluginSettings();
 		},
 	},
 };