+++ src/store/index.js
@@ -9,8 +9,8 @@ import thumbnailConfig from "../components/Thumbnail/store/thumbnail";
 import document from "../components/AppElements/store/emailDocument";
 import menu from "../components/AppElements/store/menuData";
 import sdkConfig from "../components/Dashboard/store/sdkConfig";
+import importPluginConfig from "../components/HtmlImportPlugin/store/htmlImportPlugin";
 import generatorConfig from "../components/HtmlGenerator/store/htmlGenerator";
-import importConfig from "../components/HtmlImport/store/htmlImport";
 import actions from "./actions";
 import mutations from "./mutations";
 import getters from "./getters";
@@ -30,7 +30,7 @@ export default new Vuex.Store({
 		sdkConfig,
 		menu,
 		generatorConfig,
-		importConfig,
+		importPluginConfig,
 	},
 	state: getDefaultState(),
 	mutations,